# docs/swagger.yaml
basePath: /
definitions:
  github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.SyncUserNoteWithID:
    properties:
      archived_at:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      pinned_at:
        type: string
      starred_at:
        type: string
      sync_note_tags:
        items:
          $ref: '#/definitions/github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserNoteTag'
        type: array
      sync_note_versions:
        items:
          $ref: '#/definitions/github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserNoteVersionWithID'
        type: array
      title:
        type: string
      trashed_at:
        type: string
      updated_at:
        type: string
    type: object
  github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserNoteTag:
    properties:
      assigned_at:
        type: string
      deleted_at:
        type: string
      tag_id:
        type: integer
    type: object
  github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserNoteTagWithID:
    properties:
      assigned_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      tag_id:
        type: integer
    type: object
  github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserNoteVersionWithID:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      encrypted_content:
        type: string
      id:
        type: integer
      note_id:
        type: integer
    type: object
...
  github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserTag:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_router_api_v1_note_tags.AddUserNoteTagsRequest:
    properties:
      note_id:
        type: integer
      tags_id:
        type: string
    type: object
  internal_router_api_v1_note_tags.ListUserNoteTagsRequest:
    properties:
      note_id:
        type: integer
    type: object
  internal_router_api_v1_note_tags.ListUserNoteTagsResponseBody:
    properties:
      data:
        $ref: '#/definitions/internal_router_api_v1_note_tags.ListUserNoteTagsResponseData'
      status:
        type: string
    type: object
  internal_router_api_v1_note_tags.ListUserNoteTagsResponseData:
    properties:
      note_tags:
        items:
          $ref: '#/definitions/github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserNoteTagWithID'
        type: array
    type: object
  internal_router_api_v1_note_tags.RemoveUserNoteTagsRequest:
    properties:
      note_id:
        type: integer
      tags_id:
        type: string
    type: object
...
  internal_router_api_v1_notes.ListUserNotesResponseBody:
    properties:
      data:
        $ref: '#/definitions/internal_router_api_v1_notes.ListUserNotesResponseData'
      status:
        type: string
    type: object
  internal_router_api_v1_notes.ListUserNotesResponseData:
    properties:
      notes:
        items:
          type: integer
        type: array
    type: object
  internal_router_api_v1_notes.SyncUserNotesByLastSyncedAtResponseBody:
    properties:
      data:
        $ref: '#/definitions/internal_router_api_v1_notes.SyncUserNotesByLastSyncedAtResponseData'
      status:
        type: string
    type: object
  internal_router_api_v1_notes.SyncUserNotesByLastSyncedAtResponseData:
    properties:
      sync_notes:
        items:
          $ref: '#/definitions/github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.SyncUserNoteWithID'
        type: array
    type: object
  internal_router_api_v1_tag.CreateUserTagRequest:
    properties:
      name:
        type: string
    type: object
  internal_router_api_v1_tag.DeleteUserTagRequest:
    properties:
      tag_id:
        type: integer
    type: object
  internal_router_api_v1_tag.GetUserTagByIDRequest:
    properties:
      tag_id:
        type: integer
    type: object
  internal_router_api_v1_tag.GetUserTagByIDResponseBody:
    properties:
      data:
        $ref: '#/definitions/internal_router_api_v1_tag.GetUserTagByIDResponseData'
      status:
        type: string
    type: object
  internal_router_api_v1_tag.GetUserTagByIDResponseData:
    properties:
      tag:
        $ref: '#/definitions/github_com_ralvarezdev_uru-frameworks-secure-notes-api_internal_databases_postgres_model.UserTag'
    type: object
  internal_router_api_v1_tag.UpdateUserTagRequest:
    properties:
      name:
        type: string
      tag_id:
        type: integer
    type: object
...
  response.JSendErrorBody:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
  response.JSendFailBody:
    properties:
      code:
        type: string
      data: {}
      status:
        type: string
    type: object
  response.JSendSuccessBody:
    properties:
      data: {}
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is the REST API for the Secure Notes application.
  license:
    name: GPL-3.0
    url: http://www.gnu.org/licenses/gpl-3.0.html
  title: Secure Notes REST API
  version: "1.0"
paths:
...
  /api/v1/note/tags:
    delete:
      consumes:
      - application/json
      description: Removes user note tags
      parameters:
      - description: Remove User UserNote Tags Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_note_tags.RemoveUserNoteTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Remove user note tags
      tags:
      - api v1 note
    get:
      consumes:
      - application/json
      description: List user note tags
      parameters:
      - description: List User UserNote Tags Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_note_tags.ListUserNoteTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_router_api_v1_note_tags.ListUserNoteTagsResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: List user note tags
      tags:
      - api v1 note
    patch:
      consumes:
      - application/json
      description: Adds user note tags
      parameters:
      - description: Add User UserNote Tags Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_note_tags.AddUserNoteTagsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Add user note tags
      tags:
      - api v1 note
...
  /api/v1/notes:
    get:
      consumes:
      - application/json
      description: List user notes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_router_api_v1_notes.ListUserNotesResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: List user notes
      tags:
      - api v1 notes
  /api/v1/notes/sync:
    post:
      consumes:
      - application/json
      description: Sync user notes by last synced at timestamp
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_router_api_v1_notes.SyncUserNotesByLastSyncedAtResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Sync user notes by last synced at timestamp
      tags:
      - api v1 notes
...
  /api/v1/tag:
    delete:
      consumes:
      - application/json
      description: Deletes a user tag
      parameters:
      - description: Delete User Tag Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_tag.DeleteUserTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Delete a user tag
      tags:
      - api v1 tag
    get:
      consumes:
      - application/json
      description: Gets a user tag by tag ID
      parameters:
      - description: Get User Tag By ID Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_tag.GetUserTagByIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_router_api_v1_tag.GetUserTagByIDResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Get a user tag by tag ID
      tags:
      - api v1 tag
    post:
      consumes:
      - application/json
      description: Creates a user tag
      parameters:
      - description: Create User Tag Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_tag.CreateUserTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Create a user tag
      tags:
      - api v1 tag
    put:
      consumes:
      - application/json
      description: Updates a user tag
      parameters:
      - description: Update User Tag Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_router_api_v1_tag.UpdateUserTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSendSuccessBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSendFailBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSendErrorBody'
      security:
      - CookieAuth: []
      summary: Update a user tag
      tags:
      - api v1 tag
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"